opensearch-client/requirements.txt
opensearch-py==2.0.0
mysql-connector-python==8.0.30
python-dateutil==2.8.2